update: result truncation in string_to_natural_int

+++ src/adt/string.h
@@ -65,6 +65,7 @@ void insert_char_last(char c, string *s);
 /* mengonversi string numerik menjadi integer bilangan cacah */
 /* return -1 jika tidak numerik */
 /* return -2 jika neff(s) == 0 */
-long string_to_natural_int(string s);
+/* truncate hasil jika digit lebih dari 9*/
+int string_to_natural_int(string s);
 
 #endif
